requirements.txt
Flask==3.0.0
google_api_python_client==2.108.0
numpy==1.26.2
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==10.1.0
protobuf==3.19.6
gunicorn==21.2.0